requirements.txt
torch==2.2.2
torchvision==0.17.2
gradio==4.36.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.4.2
